Performance improvement / small Bugfix of the Frontend
Currently every 3.5 seconds the entire DOM is searched for elements with
a special classname to hide them, if found (after a rails flash).

This commit replaces the continious searching for the elements in an
endless loop. Acentralized but global function call with classnames and
timeout is used instead. It's called after the Rails-flashing, inline.

Bugfix: The formerly used "drop" in the hide-call is an effect of jQuery UI which is not included
in graylog.